backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Kitmatch FastAPI service"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "pandas==2.1.3",
    "numpy==1.26.4",
    "python-dotenv==1.0.0",
    "pydantic-settings==2.1.0",
    "clickhouse-driver==0.2.6",
    "tqdm==4.66.1",
    "openpyxl==3.1.2",
    "loguru==0.7.2",
    "opencv-python-headless==4.10.0.84",
    "python-multipart==0.0.9",
    "tritonclient[http]==2.49.0",
    "pillow==11.3.0",
    "pika==1.3.2",
    "requests==2.32.5",
    "pymongo==4.15.1",
    
]

[project.optional-dependencies]
dev = [
    "black==23.12.0",
    "isort==5.13.2",
    "ruff==0.1.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kitmatch/src/kitmatch"]